requirements.txt
lightgbm
pandas
py-cpuinfo
